requirements.txt
flask
flask-bcrypt
flask-cors
flask-sqlalchemy
flask-migrate
python-dotenv
requests
jwt
gunicorn
